requirements.txt
absl-py==0.15.0
apache-beam==2.37.0
astunparse==1.6.3  
avro-python3==1.10.2
Cython==0.29.28
fastavro==1.4.7
Flask==2.0.3
Flask-Cors==3.0.10
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
#importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
#keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.0
#object-detection==0.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
orjson==3.6.1
Pillow==8.4.0
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==6.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.15.0
#tensorboard==2.8.0
#tensorboard-data-server==0.6.1
#tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
#tf-models-official==2.8.0
tf-slim==1.1.0
#typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
wget==3.2
#Detectron
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cpu
torchvision==0.11.2+cpu
torchaudio==0.10.1
cloudpickle
omegaconf
pycocotools-windows
fvcore

#Yolo
seaborn>=0.11.0
youtube_dl